refactor: switch mcp siblings to self-reexport imports
