+++ packages/opencode/src/mcp/index.ts
@@ -1,3 +1,1 @@
 export * as MCP from "./mcp"
-export * as McpAuth from "./auth"
-export * as McpOAuthCallback from "./oauth-callback"
